requirements.txt

fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
cryptography
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
requests
httpx
Pillow
greenlet
minio
email-validator
google-generativeai
pgvector
python-dotenv
pydantic-settings
python-socketio
